torch.testing allclose -> close
Right now the functions have divergent names with one postfixed `_equal` and the other `_allclose`. I've opted to use `_(equal|close)` over `_all(equal|close)` think it is a reasonable assumption that all values need to be equal or close for this pass even without explicitly naming the function this way.

ghstack-source-id: 3e51d6f7ecc6d841b1dc5e1073d4587f0e855745
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54781